<feat>: rails application
generated CRUD API on a User class with first name and lastname. Added count API on user/count to count the total number of users. Added name API on user/user_id/name to show first name and last name of the user. added a gender field to user model. added validation to User model: the first name and last name field input cannot be blank, the age field should be positive integer, the gender field that can be one of 'male', 'female' and 'other'. changed the database from sqlite3 to mongodb and updated the user model. used Jbuilder to render JSON response for JSON request. Created a address field which is a hash and include keys(country, address_1, address_2,). created a shop class that has 1 to 1 relationship to user class and generated CRUD route for Shop under user.

feat